test: isolate legacy session diagnostics
